add useScript hook for loading third-party JavaScript
This adds a useScript hook which takes a src parameter, creates a script
tag, sets the `src` and `async` attributes, then appends the script to
the document head.

This also refactors the `<Wistia Video>` component to *not* check for
whether or not a given script has been loaded already. There is no need
to check, since duplicate script tags will not trigger duplicate
requests.