fix bug where Stripe would not be initialized on second mount
This fixes a bug in the useStripe hook, where Stripe would only be
initialized when the Stripe script loaded. This was a problem because
the useScript hook only loads a given script once, so on subsequent
mounts, the script tag would already be in the document head, and thus
would not trigger the load event.

This adds a loaded state variable to the useScript hook which is set to
truthy if:

- The script is already in the document head
- The script is added to the head and its load event fires

The useScript hook returns [loaded], so it can be consumed by container
components.